Cleanup layer hierarchy
Get rid of AbstractLayer and LayerGroup models and remain on a simple layer model with source and appearance.
