--- src/shogun2-core/src/main/java/de/terrestris/shogun2/model/layer/AbstractLayer.java
@@ -1,138 +1,0 @@
-package de.terrestris.shogun2.model.layer;
-
-import javax.persistence.Entity;
-import javax.persistence.Inheritance;
-import javax.persistence.InheritanceType;
-
-import org.apache.commons.lang3.builder.EqualsBuilder;
-import org.apache.commons.lang3.builder.HashCodeBuilder;
-import org.apache.commons.lang3.builder.ToStringBuilder;
-import org.apache.commons.lang3.builder.ToStringStyle;
-
-import de.terrestris.shogun2.model.PersistentObject;
-
-/**
- *
- * Representation of a layer which consists a corresponding data source
- * and an appearance
- *
- * @author Kai Volland
- * @author Nils Bühner
- * @author terrestris GmbH & Co. KG
- *
- */
-@Entity
-@Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
-public abstract class AbstractLayer extends PersistentObject {
-
-	/**
-	 *
-	 */
-	private static final long serialVersionUID = 1L;
-
-	/**
-	 *
-	 */
-	private String name;
-
-	/**
-	 *
-	 */
-	private String type;
-
-	/**
-	 *
-	 */
-	public AbstractLayer() {
-		super();
-	}
-
-	/**
-	 * @param name Layer name
-	 * @param type Layer type
-	 * @param source The data source of the layer
-	 * @param appearance The appearance configuration of the layer
-	 */
-	public AbstractLayer(String name, String type) {
-		super();
-		this.name = name;
-		this.type = type;
-	}
-
-	/**
-	 * @return the name
-	 */
-	public String getName() {
-		return name;
-	}
-
-	/**
-	 * @param name the name to set
-	 */
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	/**
-	 * @return the type
-	 */
-	public String getType() {
-		return type;
-	}
-
-	/**
-	 * @param type the type to set
-	 */
-	public void setType(String type) {
-		this.type = type;
-	}
-
-
-	/**
-	 * @see java.lang.Object#hashCode()
-	 *
-	 *      According to
-	 *      http://stackoverflow.com/questions/27581/overriding-equals
-	 *      -and-hashcode-in-java it is recommended only to use getter-methods
-	 *      when using ORM like Hibernate
-	 */
-	@Override
-	public int hashCode() {
-		// two randomly chosen prime numbers
-		return new HashCodeBuilder(29, 13).
-				appendSuper(super.hashCode()).
-				append(getName()).
-				append(getType()).
-				toHashCode();
-	}
-
-	/**
-	 * @see java.lang.Object#equals(java.lang.Object)
-	 *
-	 *      According to
-	 *      http://stackoverflow.com/questions/27581/overriding-equals
-	 *      -and-hashcode-in-java it is recommended only to use getter-methods
-	 *      when using ORM like Hibernate
-	 */
-	@Override
-	public boolean equals(Object obj) {
-		if (!(obj instanceof AbstractLayer))
-			return false;
-		AbstractLayer other = (AbstractLayer) obj;
-
-		return new EqualsBuilder().
-				appendSuper(super.equals(other)).
-				append(getName(), other.getName()).
-				append(getType(), other.getType()).
-				isEquals();
-	}
-
-	/* (non-Javadoc)
-	 * @see java.lang.Object#toString()
-	 */
-	@Override
-	public String toString() {
-		return ToStringBuilder.reflectionToString(this, ToStringStyle.DEFAULT_STYLE);
-	}
-
-}
