+++ src/shogun2-core/src/main/java/de/terrestris/shogun2/model/layer/appearance/LayerAppearance.java
@@ -9,7 +9,6 @@ import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.apache.commons.lang3.builder.ToStringStyle;
 
 import de.terrestris.shogun2.model.PersistentObject;
-import de.terrestris.shogun2.model.layer.Layer;
 
 /**
  *
