--- src/shogun2-core/src/main/java/de/terrestris/shogun2/rest/LayerGroupRestController.java
@@ -1,49 +1,0 @@
-package de.terrestris.shogun2.rest;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import de.terrestris.shogun2.dao.LayerGroupDao;
-import de.terrestris.shogun2.model.layer.LayerGroup;
-import de.terrestris.shogun2.service.LayerGroupService;
-
-/**
- * @author Kai Volland
- * @author Nils Bühner
- *
- */
-@RestController
-@RequestMapping("/layergroups")
-public class LayerGroupRestController<E extends LayerGroup, D extends LayerGroupDao<E>, S extends LayerGroupService<E, D>>
-		extends AbstractRestController<E, D, S> {
-
-	/**
-	 * Default constructor, which calls the type-constructor
-	 */
-	@SuppressWarnings("unchecked")
-	public LayerGroupRestController() {
-		this((Class<E>) LayerGroup.class);
-	}
-
-	/**
-	 * Constructor that sets the concrete entity class for the controller.
-	 * Subclasses MUST call this constructor.
-	 */
-	protected LayerGroupRestController(Class<E> entityClass) {
-		super(entityClass);
-	}
-
-	/**
-	 * We have to use {@link Qualifier} to define the correct service here.
-	 * Otherwise, spring can not decide which service has to be autowired here
-	 * as there are multiple candidates.
-	 */
-	@Override
-	@Autowired
-	@Qualifier("layerGroupService")
-	public void setService(S service) {
-		this.service = service;
-	}
-}
